+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This document highlights high-level changes made to this program.
 
 ## Unreleased
 + Use enums for `data_type`. (#61)
++ Use an ordered dict internally for `DiscreteLegend`.
++ Add a `discrete_legend_colors` arg to the WaferMapPanel constructor.
+
 
 ## 1.0.25 / 2018-10-10
 + Fixed rendering in README.rst (for PyPI's `long_description`). Again.